Refactor code by spliting interface and implementation
