+++ demo/rails/Gemfile
@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 
 gem 'dotenv-rails', groups: [:development, :test]
-gem 'rails', '4.2.5'
+gem 'rails', '~> 4.2.11.1'
 gem 'sqlite3'
 gem 'recaptcha', require: 'recaptcha/rails', path: '../..'
