+++ lib/recaptcha/configuration.rb
@@ -30,12 +30,17 @@ module Recaptcha
   #   end
   #
   class Configuration
+    DEFAULTS = {
+      'server_url' => 'https://www.google.com/recaptcha/api.js',
+      'verify_url' => 'https://www.google.com/recaptcha/api/siteverify'
+    }.freeze
+
     attr_accessor :skip_verify_env, :secret_key, :site_key, :proxy, :handle_timeouts_gracefully, :hostname
     attr_writer :api_server_url, :verify_url
 
     def initialize #:nodoc:
       @skip_verify_env = %w[test cucumber]
-      @handle_timeouts_gracefully = HANDLE_TIMEOUTS_GRACEFULLY
+      @handle_timeouts_gracefully = true
 
       @secret_key = ENV['RECAPTCHA_SECRET_KEY']
       @site_key = ENV['RECAPTCHA_SITE_KEY']
@@ -52,11 +57,11 @@ module Recaptcha
     end
 
     def api_server_url
-      @api_server_url || CONFIG.fetch('server_url')
+      @api_server_url || DEFAULTS.fetch('server_url')
     end
 
     def verify_url
-      @verify_url || CONFIG.fetch('verify_url')
+      @verify_url || DEFAULTS.fetch('verify_url')
     end
   end
 end
