+++ test/client_helper_test.rb
@@ -2,8 +2,8 @@ require_relative 'helper'
 require 'active_support/core_ext/object/blank'
 require 'active_support/core_ext/hash'
 
-describe Recaptcha::ClientHelper do
-  include Recaptcha::ClientHelper
+describe 'View helpers' do
+  include Recaptcha::Adapters::View
 
   it "uses ssl" do
     recaptcha_tags.must_include "\"#{Recaptcha.configuration.api_server_url}\""
