+++ test/verify_test.rb
@@ -1,6 +1,6 @@
 require_relative 'helper'
 
-describe Recaptcha::Verify do
+describe 'controller helpers' do
   before do
     @controller = TestController.new
     @controller.request = stub(remote_ip: "1.1.1.1", format: :html)
@@ -255,7 +255,7 @@ describe Recaptcha::Verify do
   private
 
   class TestController
-    include Recaptcha::Verify
+    include Recaptcha::Adapters::Controller
     attr_accessor :request, :params, :flash
 
     def initialize
